UI 操作逻辑优化
UI 逻辑: 添加滑动删除/撤销，调整编辑按钮至跟踪项详情界面，套用 MD 主题

+++ app/src/main/kotlin/net/xzos/upgradeAll/server/app/manager/module/App.kt
@@ -5,7 +5,7 @@ import net.xzos.upgradeAll.utils.VersionChecker
 import org.litepal.LitePal
 import org.litepal.extension.find
 
-data class App(private val appDatabaseId: Int) {
+data class App(private val appDatabaseId: Long) {
     val updater: Updater = Updater(appDatabaseId)
 
     val isLatest: Boolean
